Add task state/message logging.
In addition to the UI, this PR adds a simple logging based on simple
diffing of task states. This happens in lockstep with polling interval
(thus, this doesn't emit any transient changes while the tasks are
running), but should provide a reasonable logging output in cases where
the UI is not available (or timestamps are needed).

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>