Refactor NewShellTask signature
This change will keep the existing code functional as it doesn't
change the calling sites while adding an extra layer of "security"
whereas you are required to pass a command to this task. Previously
you could call this like `t := NewShellTask("no command")` and the
underlying exec.Command will be created without an actual command to run.

Signed-off-by: Leandro López (inkel) <inkel.ar@gmail.com>