Implement custom Poll() for NestedTask.
Up until now, `NestedTask`'s Poll() only performed a simple scheduling
function, not allowing for any customization. This PR adds a support for
custom callbacks that are (to some extent) able to modify the children
of NestedTasks, making it possible to update the children in the
real-time to reflect the real-world situations.

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>