Fix oldState handling (stop using pointers to interfaces).
`oldState` in NestedTask was originally meant to be used for indicating
that a task state change happened and a Poll() is necessary. However, by
using pointers to interfaces instead of interface itself, this might've
inadvertedly been doing some weird things. Let's keep things simple by
just using the interfaces.

This was discovered when adding tests exercising the `oldState` Poll()
triggering.

Signed-off-by: Milan Plzik <milan.plzik@grafana.com>